new topo files as the projection attributes changed with a newer version of geopandas, moved veg topo tests to new file
