+++ tests/Lakes/gold/landfire_140/gold_config.ini
@@ -1,11 +1,11 @@
 [generate_topo]
 
-basin_shapefile:        ./basin_outline.shp
-dem_file:               ../data/dem_epsg_32611_100m.tif
+basin_shapefile:        ../basin_outline.shp
+dem_file:               ../../data/dem_epsg_32611_100m.tif
 output_folder:          ./
 basin_name:             Lakes
 
-vegetation_folder:      ../../../landfire/landfire_140
+vegetation_folder:      ../../../../landfire/landfire_140
 vegetation_dataset:     landfire_1.4.0
 
 pad_domain:             5, 5, 5, 5
